requirements.txt
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.11.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.1
click==8.1.8
durationpy==0.9
fastapi==0.116.1
Flask==3.1.0
google-auth==2.39.0
graphviz==0.20.3
h11==0.16.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
kubernetes==32.0.1
MarkupSafe==3.0.2
oauthlib==3.2.2
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.47.1
toml==0.10.2
typing-inspection==0.4.1
typing_extensions==4.14.1
tzlocal==5.3.1
urllib3==2.4.0
uvicorn==0.35.0
websocket-client==1.8.0
Werkzeug==3.1.3
